# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  model.Machine:
    properties:
      cpuname:
        type: string
      cputemp:
        type: number
      createdAt:
        type: string
      deletedAt:
        type: string
      gpuname:
        type: string
      gputemp:
        type: number
      groupID:
        type: string
      id:
        type: string
      ipaddress:
        type: string
      isActive:
        type: boolean
      lastHeartbeat:
        type: string
      macAddress:
        type: string
      machineCode:
        type: string
      osinfo:
        type: string
      ramgb:
        type: integer
      status:
        type: string
      storageGB:
        type: integer
      storeID:
        type: string
      updatedAt:
        type: string
    type: object
  model.MachineAsset:
    properties:
      assetType:
        type: string
      brand:
        type: string
      checkPhotos:
        items:
          type: string
        type: array
      checkedAt:
        type: string
      checkedBy:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      machineID:
        type: string
      model:
        type: string
      notes:
        type: string
      serial:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  model.MachineGroup:
    properties:
      color:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      sortOrder:
        type: integer
      storeID:
        type: string
    type: object
...
  response.PaginatedData:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.CreateMachineAssetRequest:
    properties:
      asset_type:
        type: string
      brand:
        type: string
      machine_id:
        type: string
      model:
        type: string
      notes:
        type: string
      serial:
        type: string
      status:
        type: string
    required:
    - asset_type
    - machine_id
    type: object
  service.CreateMachineGroupRequest:
    properties:
      color:
        type: string
      description:
        type: string
      name:
        type: string
      sort_order:
        type: integer
    required:
    - name
    type: object
...
  service.CreateMachineRequest:
    properties:
      cpu_name:
        type: string
      gpu_name:
        type: string
      group_id:
        type: string
      machine_code:
        type: string
      os_info:
        type: string
      ram_gb:
        type: integer
      storage_gb:
        type: integer
    required:
    - machine_code
    type: object
...
  service.UpdateMachineAssetRequest:
    properties:
      asset_type:
        type: string
      brand:
        type: string
      model:
        type: string
      notes:
        type: string
      serial:
        type: string
      status:
        type: string
    type: object
  service.UpdateMachineGroupRequest:
    properties:
      color:
        type: string
      description:
        type: string
      name:
        type: string
      sort_order:
        type: integer
    type: object
...
  service.UpdateMachineRequest:
    properties:
      cpu_name:
        type: string
      gpu_name:
        type: string
      group_id:
        type: string
      os_info:
        type: string
      ram_gb:
        type: integer
      status:
        type: string
      storage_gb:
        type: integer
    type: object
...
host: localhost:20800
info:
  contact:
    email: support@vnet.net
    name: VNET Support
  description: Hệ thống Quản lý Phòng Game Toàn Diện
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://vnet.net/terms
  title: VNET Core API
  version: "1.0"
paths:
...
  /api/machine-assets:
    get:
      consumes:
      - application/json
      description: Get machine assets, optionally filtered by machine
      parameters:
      - description: Filter by machine ID
        in: query
        name: machine_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.MachineAsset'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List Machine Assets
      tags:
      - MachineAssets
    post:
      consumes:
      - application/json
      description: Create a new machine asset record
      parameters:
      - description: Machine asset data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateMachineAssetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.MachineAsset'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create Machine Asset
      tags:
      - MachineAssets
  /api/machine-assets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a machine asset by its ID
      parameters:
      - description: Machine Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete Machine Asset
      tags:
      - MachineAssets
    put:
      consumes:
      - application/json
      description: Update an existing machine asset
      parameters:
      - description: Machine Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Machine asset update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UpdateMachineAssetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.MachineAsset'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update Machine Asset
      tags:
      - MachineAssets
  /api/machine-groups:
    get:
      consumes:
      - application/json
      description: Get all machine groups for the current store
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.MachineGroup'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List Machine Groups
      tags:
      - MachineGroups
    post:
      consumes:
      - application/json
      description: Create a new machine group
      parameters:
      - description: Machine group data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateMachineGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.MachineGroup'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create Machine Group
      tags:
      - MachineGroups
  /api/machine-groups/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a machine group by its ID
      parameters:
      - description: Machine Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete Machine Group
      tags:
      - MachineGroups
    put:
      consumes:
      - application/json
      description: Update an existing machine group
      parameters:
      - description: Machine Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Machine group update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UpdateMachineGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.MachineGroup'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update Machine Group
      tags:
      - MachineGroups
...
  /api/machines:
    get:
      consumes:
      - application/json
      description: Get paginated list of machines for the current store
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      - description: Search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PaginatedData'
                  - properties:
                      data:
                        items:
                          $ref: '#/definitions/model.Machine'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List Machines
      tags:
      - Machines
    post:
      consumes:
      - application/json
      description: Create a new machine
      parameters:
      - description: Machine data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateMachineRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Machine'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create Machine
      tags:
      - Machines
  /api/machines/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a machine by its ID
      parameters:
      - description: Machine ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete Machine
      tags:
      - Machines
    get:
      consumes:
      - application/json
      description: Get a single machine by its ID
      parameters:
      - description: Machine ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Machine'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Machine by ID
      tags:
      - Machines
    put:
      consumes:
      - application/json
      description: Update an existing machine
      parameters:
      - description: Machine ID
        in: path
        name: id
        required: true
        type: string
      - description: Machine update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UpdateMachineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Machine'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update Machine
      tags:
      - Machines
...
  /api/machines/{id}/hardware:
    get:
      consumes:
      - application/json
      description: Get paginated hardware telemetry history for a machine
      parameters:
      - description: Machine ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PaginatedData'
                  - properties:
                      data:
                        items:
                          $ref: '#/definitions/model.MachineHardwareSnapshot'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Hardware History
      tags:
      - Machines
  /api/machines/{id}/heartbeat:
    post:
      consumes:
      - application/json
      description: Send heartbeat from a machine with hardware telemetry
      parameters:
      - description: Machine ID
        in: path
        name: id
        required: true
        type: string
      - description: Heartbeat data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.HeartbeatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Machine Heartbeat
      tags:
      - Machines
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"